Move PaymentSelection to shared components
