+++ src/shared/assets/images/selected_desktop.png
